refactor: rename util/csv package to util/export
The package exports both csv and xlsx, so csvutil.NewLocalizedXlsx read
wrong. Rename package to export, NewLocalizedWriter to NewLocalizedCsv
for symmetry with NewLocalizedXlsx, and drop the unreachable error guard
in xlsxWriter.Write.
